Don't hard-code sequences
Instead keep track of the next expected sequence and validate that we're
in agreement with what the server sends. This does require resetting
back to zero before each command, but that also gives us a chance to
double check that we don't have any unexpected data lurking in our
buffers.