Add gtid support
This gets the gtid tests passing. It also moves the status update stuff
so that it catches *every* OK packet (there may have been some missing
in some of the other commands like change_db). This does complicate
Connection#read_packet a bit, but it also makes things simpler for each
caller.

We have to check the length to determine if we are looking at an OK
packet because a row with an empty value could also start with a 0 (i.e.
a length encoded str where the length is 0).